- Merged dev branch (docs migration to docs-mslearn/)
- Created separate Virtual machine scale sets section per flanakin's review
- Updated changelog to reflect documentation improvements

Resolves conflicts from docs migration (#1353)
Addresses review comment #11 in PR #1088

+++ docs/deploy/optimization-engine/latest/runbooks/data-collection/Ingest-OptimizationCSVExportsToLogAnalytics.ps1
@@ -54,12 +54,12 @@ $LogAnalyticsIngestControlTable = "LogAnalyticsIngestControl"
 "Logging in to Azure with $authenticationOption..."
 
 switch ($authenticationOption) {
-    "UserAssignedManagedIdentity" { 
+    "UserAssignedManagedIdentity" {
         Connect-AzAccount -Identity -EnvironmentName $cloudEnvironment -AccountId $uamiClientID
         break
     }
     Default { #ManagedIdentity
-        Connect-AzAccount -Identity -EnvironmentName $cloudEnvironment 
+        Connect-AzAccount -Identity -EnvironmentName $cloudEnvironment
         break
     }
 }
@@ -95,7 +95,7 @@ Function Post-OMSData($workspaceId, $sharedKey, $body, $logType, $TimeStampField
         -method $method `
         -contentType $contentType `
         -resource $resource
-    
+
     $uri = "https://" + $workspaceId + ".ods.opinsights.azure.com" + $resource + "?api-version=2016-04-01"
     if ($AzureEnvironment -eq "AzureChinaCloud")
     {
@@ -122,7 +122,7 @@ Function Post-OMSData($workspaceId, $sharedKey, $body, $logType, $TimeStampField
         $response = Invoke-WebRequest -Uri $uri -Method POST  -ContentType $contentType -Headers $OMSheaders -Body $body -UseBasicParsing -TimeoutSec 1000
     }
     catch {
-        if ($_.Exception.Response.StatusCode.Value__ -eq 401) {            
+        if ($_.Exception.Response.StatusCode.Value__ -eq 401) {
             "REAUTHENTICATING"
 
             $response = Invoke-WebRequest -Uri $uri -Method POST  -ContentType $contentType -Headers $OMSheaders -Body $body -UseBasicParsing -TimeoutSec 1000
@@ -133,7 +133,7 @@ Function Post-OMSData($workspaceId, $sharedKey, $body, $logType, $TimeStampField
         }
     }
 
-    return $response.StatusCode    
+    return $response.StatusCode
 }
 #endregion Functions
 
@@ -163,14 +163,14 @@ do {
     $tries++
     try {
         $dbToken = Get-AzAccessToken -ResourceUrl "https://$azureSqlDomain/"
-        $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;") 
+        $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;")
         $Conn.AccessToken = $dbToken.Token
-        $Conn.Open() 
+        $Conn.Open()
         $Cmd=new-object system.Data.SqlClient.SqlCommand
         $Cmd.Connection = $Conn
         $Cmd.CommandTimeout = $SqlTimeout
         $Cmd.CommandText = "SELECT * FROM [dbo].[$LogAnalyticsIngestControlTable] WHERE StorageContainerName = '$storageAccountSinkContainer'"
-    
+
         $sqlAdapter = New-Object System.Data.SqlClient.SqlDataAdapter
         $sqlAdapter.SelectCommand = $Cmd
         $controlRows = New-Object System.Data.DataTable
@@ -181,7 +181,7 @@ do {
         Write-Output "Failed to contact SQL at try $tries."
         Write-Output $Error[0]
         Start-Sleep -Seconds ($tries * 20)
-    }    
+    }
 } while (-not($connectionSuccess) -and $tries -lt 3)
 
 if (-not($connectionSuccess))
@@ -189,8 +189,8 @@ if (-not($connectionSuccess))
     throw "Could not establish connection to SQL."
 }
 
-$Conn.Close()    
-$Conn.Dispose()            
+$Conn.Close()
+$Conn.Dispose()
 
 if ($controlRows.Count -eq 0 -or -not($controlRows[0].LastProcessedDateTime))
 {
@@ -224,7 +224,7 @@ Write-Output "Found $($unprocessedBlobs.Count) new blobs to process..."
 
 foreach ($blob in $unprocessedBlobs) {
     $newProcessedTime = $blob.LastModified.UtcDateTime.ToString("yyyy'-'MM'-'dd'T'HH':'mm':'ss'.'fff'Z'")
-    Write-Output "About to process $($blob.Name)..."
+    Write-Output "About to process $($blob.Name) ($($blob.Length) bytes)..."
     $blobFilePath = "$env:TEMP\$($blob.Name)"
     Get-AzStorageBlobContent -CloudBlob $blob.ICloudBlob -Context $saCtx -Force -Destination $blobFilePath | Out-Null
 
@@ -234,7 +234,7 @@ foreach ($blob in $unprocessedBlobs) {
     $lineCounter = 0
     $chunkLines = @()
 
-    while ($r.Peek() -ge 0) 
+    while ($r.Peek() -ge 0)
     {
         $line = $r.ReadLine()
         if ($lineCounter -eq 0)
@@ -244,7 +244,7 @@ foreach ($blob in $unprocessedBlobs) {
         }
         else
         {
-            $linesProcessed++    
+            $linesProcessed++
         }
         if ($lastProcessedLine -lt $linesProcessed -and $lineCounter -gt 0)
         {
@@ -255,44 +255,53 @@ foreach ($blob in $unprocessedBlobs) {
             $csvObject = $chunkLines | ConvertFrom-Csv
             $jsonObject = ConvertTo-Json -InputObject $csvObject
 
-            $res = Post-OMSData -workspaceId $workspaceId -sharedKey $sharedKey -body ([System.Text.Encoding]::UTF8.GetBytes($jsonObject)) -logType $logname -TimeStampField "Timestamp" -AzureEnvironment $cloudEnvironment
-            if ($res -ge 200 -and $res -lt 300) 
+            if ($null -ne $jsonObject)
             {
-                Write-Output "Succesfully uploaded $lineCounter $LogAnalyticsSuffix rows to Log Analytics"    
-                if ($r.Peek() -lt 0) {
-                    $lastProcessedLine = -1    
+                $res = Post-OMSData -workspaceId $workspaceId -sharedKey $sharedKey -body ([System.Text.Encoding]::UTF8.GetBytes($jsonObject)) -logType $logname -TimeStampField "Timestamp" -AzureEnvironment $cloudEnvironment
+
+                if ($res -ge 200 -and $res -lt 300)
+                {
+                    Write-Output "Succesfully uploaded $lineCounter $LogAnalyticsSuffix rows to Log Analytics"
                 }
-                else {
-                    $lastProcessedLine = $linesProcessed - 1   
+                else
+                {
+                    Write-Warning "Failed to upload $lineCounter $LogAnalyticsSuffix rows. Error code: $res"
+                    $r.Dispose()
+                    Remove-Item -Path $blobFilePath -Force
+                    throw
                 }
-                
-                $updatedLastProcessedLine = $lastProcessedLine
-                $updatedLastProcessedDateTime = $lastProcessedDateTime
-                if ($r.Peek() -lt 0) {
-                    $updatedLastProcessedDateTime = $newProcessedTime
-                }
-                $lastProcessedDateTime = $updatedLastProcessedDateTime
-                Write-Output "Updating last processed time / line to $($updatedLastProcessedDateTime) / $updatedLastProcessedLine"
-                $sqlStatement = "UPDATE [$LogAnalyticsIngestControlTable] SET LastProcessedLine = $updatedLastProcessedLine, LastProcessedDateTime = '$updatedLastProcessedDateTime' WHERE StorageContainerName = '$storageAccountSinkContainer'"
-                $dbToken = Get-AzAccessToken -ResourceUrl "https://$azureSqlDomain/"
-                $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;") 
-                $Conn.AccessToken = $dbToken.Token
-                $Conn.Open() 
-                $Cmd=new-object system.Data.SqlClient.SqlCommand
-                $Cmd.Connection = $Conn
-                $Cmd.CommandText = $sqlStatement
-                $Cmd.CommandTimeout = $SqlTimeout
-                $Cmd.ExecuteReader()
-                $Conn.Close()    
-                $Conn.Dispose()            
             }
-            else 
+            else
             {
-                Write-Warning "Failed to upload $lineCounter $LogAnalyticsSuffix rows. Error code: $res"
-                $r.Dispose()
-                Remove-Item -Path $blobFilePath -Force
-                throw
+                Write-Warning "Skipped uploading $lineCounter $LogAnalyticsSuffix rows. Null JSON object."
             }
+
+            if ($r.Peek() -lt 0) {
+                $lastProcessedLine = -1
+            }
+            else {
+                $lastProcessedLine = $linesProcessed - 1
+            }
+
+            $updatedLastProcessedLine = $lastProcessedLine
+            $updatedLastProcessedDateTime = $lastProcessedDateTime
+            if ($r.Peek() -lt 0) {
+                $updatedLastProcessedDateTime = $newProcessedTime
+            }
+            $lastProcessedDateTime = $updatedLastProcessedDateTime
+            Write-Output "Updating last processed time / line to $($updatedLastProcessedDateTime) / $updatedLastProcessedLine"
+            $sqlStatement = "UPDATE [$LogAnalyticsIngestControlTable] SET LastProcessedLine = $updatedLastProcessedLine, LastProcessedDateTime = '$updatedLastProcessedDateTime' WHERE StorageContainerName = '$storageAccountSinkContainer'"
+            $dbToken = Get-AzAccessToken -ResourceUrl "https://$azureSqlDomain/"
+            $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;")
+            $Conn.AccessToken = $dbToken.Token
+            $Conn.Open()
+            $Cmd=new-object system.Data.SqlClient.SqlCommand
+            $Cmd.Connection = $Conn
+            $Cmd.CommandText = $sqlStatement
+            $Cmd.CommandTimeout = $SqlTimeout
+            $Cmd.ExecuteReader()
+            $Conn.Close()
+            $Conn.Dispose()
 
             $chunkLines = @()
             $chunkLines += $header
@@ -301,34 +310,34 @@ foreach ($blob in $unprocessedBlobs) {
         else
         {
             $lineCounter++
-        }        
+        }
     }
     $r.Dispose()
 
     if ($linesProcessed -eq 0)
     {
         Write-Output "No rows found"
-        $updatedLastProcessedLine = -1 
+        $updatedLastProcessedLine = -1
         $updatedLastProcessedDateTime = $newProcessedTime
         Write-Output "Updating last processed time / line to $($updatedLastProcessedDateTime) / $updatedLastProcessedLine"
         $sqlStatement = "UPDATE [$LogAnalyticsIngestControlTable] SET LastProcessedLine = $updatedLastProcessedLine, LastProcessedDateTime = '$updatedLastProcessedDateTime' WHERE StorageContainerName = '$storageAccountSinkContainer'"
         $dbToken = Get-AzAccessToken -ResourceUrl "https://$azureSqlDomain/"
-        $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;") 
+        $Conn = New-Object System.Data.SqlClient.SqlConnection("Server=tcp:$sqlserver,1433;Database=$sqldatabase;Encrypt=True;Connection Timeout=$SqlTimeout;")
         $Conn.AccessToken = $dbToken.Token
-        $Conn.Open() 
+        $Conn.Open()
         $Cmd=new-object system.Data.SqlClient.SqlCommand
         $Cmd.Connection = $Conn
         $Cmd.CommandText = $sqlStatement
         $Cmd.CommandTimeout = $SqlTimeout
         $Cmd.ExecuteReader()
-        $Conn.Close()    
-        $Conn.Dispose()            
+        $Conn.Close()
+        $Conn.Dispose()
     }
     else
     {
-        Write-Output "Processed $linesProcessed row(s) in total."  
+        Write-Output "Processed $linesProcessed row(s) in total."
     }
-    
+
     Remove-Item -Path $blobFilePath -Force
 }
